Moved src/server/ to src/
