+++ .gitignore
@@ -1,4 +1,4 @@
-src/server/blacklist.js
+src/blacklist.js
 public/blinkies-public/*gif
 public/blinkies-public/display/*png
 assets/blinkies-frames
